// src/frontend/src/app/api/reports/swagger.ts
export enum WiState {
    Design = 0,
    Closed = 1,
    Ready = 2,
}
// ...
export class SearchTestCaseItem implements ISearchTestCaseItem {
    id!: number;
    name?: string | undefined;
    priority!: number;

    constructor(data?: ISearchTestCaseItem) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.priority = _data["priority"];
        }
    }

    static fromJS(data: any): SearchTestCaseItem {
        data = typeof data === 'object' ? data : {};
        let result = new SearchTestCaseItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["priority"] = this.priority;
        return data;
    }
}

export interface ISearchTestCaseItem {
    id: number;
    name?: string | undefined;
    priority: number;
}
// ...
export class GetSuitesTreeResponse implements IGetSuitesTreeResponse {
    suites?: SuitesTreeItem[] | undefined;

    constructor(data?: IGetSuitesTreeResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["suites"])) {
                this.suites = [] as any;
                for (let item of _data["suites"])
                    this.suites!.push(SuitesTreeItem.fromJS(item));
            }
        }
    }

    static fromJS(data: any): GetSuitesTreeResponse {
        data = typeof data === 'object' ? data : {};
        let result = new GetSuitesTreeResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.suites)) {
            data["suites"] = [];
            for (let item of this.suites)
                data["suites"].push(item.toJSON());
        }
        return data;
    }
}

export interface IGetSuitesTreeResponse {
    suites?: SuitesTreeItem[] | undefined;
}

export class SuitesTreeItem implements ISuitesTreeItem {
    id!: number;
    name?: string | undefined;
    parentId?: number | undefined;

    constructor(data?: ISuitesTreeItem) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.parentId = _data["parentId"];
        }
    }

    static fromJS(data: any): SuitesTreeItem {
        data = typeof data === 'object' ? data : {};
        let result = new SuitesTreeItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["parentId"] = this.parentId;
        return data;
    }
}

export interface ISuitesTreeItem {
    id: number;
    name?: string | undefined;
    parentId?: number | undefined;
}
// ...
export enum Outcome {
    Unknow = 0,
    Planed = 1,
    Passed = 2,
    Failed = 3,
    Blocked = 4,
    Skipped = 5,
    Paused = 6,
}
// ...
export class SuiteTestCaseItem implements ISuiteTestCaseItem {
    id!: number;
    testCaseId!: number;
    name?: string | undefined;
    order!: number;
    state!: WiState;
    priority!: number;

    constructor(data?: ISuiteTestCaseItem) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.testCaseId = _data["testCaseId"];
            this.name = _data["name"];
            this.order = _data["order"];
            this.state = _data["state"];
            this.priority = _data["priority"];
        }
    }

    static fromJS(data: any): SuiteTestCaseItem {
        data = typeof data === 'object' ? data : {};
        let result = new SuiteTestCaseItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["testCaseId"] = this.testCaseId;
        data["name"] = this.name;
        data["order"] = this.order;
        data["state"] = this.state;
        data["priority"] = this.priority;
        return data;
    }
}

export interface ISuiteTestCaseItem {
    id: number;
    testCaseId: number;
    name?: string | undefined;
    order: number;
    state: WiState;
    priority: number;
}
// ...
export class ConfigsForAssign implements IConfigsForAssign {
    id!: number;
    name?: string | undefined;
    values?: string | undefined;

    constructor(data?: IConfigsForAssign) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.values = _data["values"];
        }
    }

    static fromJS(data: any): ConfigsForAssign {
        data = typeof data === 'object' ? data : {};
        let result = new ConfigsForAssign();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["values"] = this.values;
        return data;
    }
}

export interface IConfigsForAssign {
    id: number;
    name?: string | undefined;
    values?: string | undefined;
}